+++ lib/src/collection.dart
@@ -7,12 +7,11 @@ import '../bot.dart';
 
 import 'errors.dart';
 import 'helpers.dart';
-import 'math.dart';
 import 'require.dart';
 import 'typedefs.dart';
-import 'util.dart';
 
-part 'collection/array_2d.dart';
+export 'collection/array_2d.dart';
+
 part 'collection/collection_util.dart';
 part 'collection/enumerable.dart';
 part 'collection/grouping.dart';